Move `http` into own directory and split off the asset part
